Don't report all errors on each file analyzed
It's likely the original developer assumed that the pattern given to
Node#xpath is treated relative to the node on which it's called. That
doesn't seem to be the case.

The XPath "//error" finds ALL error nodes across the entire document,
even when invoked as file.xpath("//error"). This causes them all to be
reported on every file analyzed. Besides being extremely non-sensical,
this can also be a performance nightmare.

To correct this, we instead use Node#children to get only the currently
processing file's errors. This required replacing some hash accesses
with methods and the chaining of #value on attributes.