Raise default remediation, categorize checks
Categorize checks loosely based on the CSSLint wiki[1]. The most obvious
departure was categorizing Performance rules as Bug Risk checks.

1: https://github.com/CSSLint/csslint/wiki/Rules

Default remediation was at an abysmal 500, upped it to the "trivial"
baseline for now and built a small class allowing us to easily override
these later.

I chose keyword arguments in the CheckDetails class so we can easily
create an override object for one, the other, or both of categories and
remediation points.