Bump csslint
This commit fixes the checkstyle-xml formatter to include an identifier
attribute on all errors. File read errors were previously missed and
caused a NoMethodError in the engine when an identifier attribute was
used on these errors.

The commit sha was also extracted to a variable in the Dockerfile and
added as a full commit sha. This makes for easier editing (git log or
the GitHub UI most often gives you full shas) and should keep the diff
clearer that only the sha was changed.