Rerwrite in JavaScript
We-re maintaining our own fork of CSSLink because we need to make some
changes to make the output compatible with the Spec. This makes CSSLint
updates hard and as a result the engine lags behind the upstream.

CSSLint provides APIs that we can use directly to augment output but
they're only exposed in JavaScript. This rewrite is targeted towards
using the upstream CSSLint and only maintaining the custom parts that in
the engine. This should make CSSLint updates trivial as long as the API
stays stable.

As a bonus it drops Ruby from dependencies making the overal image
smaller.